refactor: seperated commandService and session to seperate files and did some basic cleaning
